+++ src/HVDCWISE_TEA.jl
@@ -38,6 +38,9 @@ include("core/variable.jl")
 include("form/acp.jl")
 include("form/dcp.jl")
 
+include("io/build_outputs.jl")
+include("io/build_raw_inputs.jl")
+include("io/load_case.jl")
 include("io/export.jl")
 include("io/parse.jl")