+++ changelog.md
@@ -9,6 +9,13 @@ All notable changes to HVCWISE_TEA will be documented in this file.
 ### Removed
 ### Fixed
 
+## [add_tests] - 2024-05-28
+### Added
+- build_outputs_from_csv()
+- build_outputs_from_json()
+### Fixed
+- parse_data() used in load_case() was deleted in the multi-period branch
+
 ## [add_tests] - 2024-05-23
 ### Added
 - test05 (1 gen & 1 load & 1 storage) implemented & validated